back_python/requirements.txt
Flask==2.0.2
requests==2.26.0
pymongo==3.11.4
python-dotenv==0.19.0
dnspython==2.1.0
jsonschema==3.2.0